python/flask-webservices-labs/flask-graphql-labs/requirements.txt
click==6.7
Flask==0.12.2
Flask-GraphQL==1.4.1
graphql-core==2.0
graphql-server-core==1.0.dev20170322001
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pkg-resources==0.0.0
promise==2.1
Rx==1.6.0
six==1.11.0
typing==3.6.4
Werkzeug==0.14.1
